backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
chromadb
openai
PyPDF2
python-docx
requests
tqdm
numpy
beautifulsoup4
